requirements.txt
Django==1.7.4
PyYAML==3.11
django-rest-swagger==0.2.8
djangorestframework==3.0.5
docutils==0.12
pydub==0.10.0
wsgiref==0.1.2
